Thread Referral
Add the initial concept of thread (tickets & tasks) referral.

* Agent & Team
Referring a ticket to an agent or team - is just like assignment without
actually assigning. If referred agent doesn't have access to the ticket's
department then they'll only have view access.

* Department
This is like assigning a ticket to the entire department. Meaning agents who
have access to the referred to department will now see the ticket - what
they can do with the ticket will depend on the assigned role in that
particular department.

It's important to note the ticket will technically be still the
responsibility of the primary department. Transferring the ticket is the
sure way to relinquishing the responsibility.

* Auto Department Referral via email
This will happen if an email is sent to multiple departments. For example an
email with TO: support &  CC: billing will result in the ticket getting
routed to "support" department and "billing" getting a referral of the same
ticket. This allows both departments to have visibility of the ticket -
which is not possible at the moment.