access: remove several joins from queries checking access
This removes two or three joins from queries which check ticket access, such as
the queue pages, by checking the object_type and object_id directly. In
general, this helps the MySQL query optimizer have an easier time realizing
that the joins are not necessary at all. It would be nice if the ORM would
realize that a join to check the primary key value of the foreign table should
actually stop one table short in the join path.